Fix #40
