+++ app/src/main/java/com/felisreader/chapter/domain/model/Chapter.kt
@@ -2,10 +2,9 @@ package com.felisreader.chapter.domain.model
 
 import com.felisreader.core.domain.model.EntityType
 import com.felisreader.core.domain.model.Relationship
-import java.util.UUID
 
 data class Chapter(
-    val id: UUID,
+    val id: String,
     val type: EntityType,
     val attributes: ChapterAttributes,
     val relationships: List<Relationship>
