+++ app/src/main/java/com/felisreader/chapter/presentation/chapter_lector/LectorScreen.kt
@@ -7,12 +7,8 @@ import androidx.compose.foundation.gestures.*
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.ArrowForward
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.pointer.pointerInput
@@ -20,10 +16,8 @@ import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
-import com.bumptech.glide.integration.compose.GlideImage
-import com.felisreader.R
-import com.felisreader.chapter.domain.model.AggregateChapter
+import com.felisreader.chapter.presentation.chapter_lector.components.NavigationField
+import com.felisreader.chapter.presentation.chapter_lector.components.PageImage
 import com.felisreader.manga.presentation.manga_info.components.Loading
 import kotlinx.coroutines.launch
 
@@ -119,7 +113,8 @@ fun LectorContent(
                     ) {
                         Text(
                             modifier = Modifier.fillMaxWidth(),
-                            text = if (state.chapter.attributes.title.isNullOrBlank()) "Chapter ${state.chapter.attributes.chapter}" else state.chapter.attributes.title,
+                            text = if (state.chapter.attributes.title.isNullOrBlank()) "Chapter ${state.chapter.attributes.chapter}"
+                                    else state.chapter.attributes.title,
                             style = MaterialTheme.typography.titleMedium,
                             textAlign = TextAlign.Center
                         )
@@ -153,62 +148,6 @@ fun LectorContent(
             }
         } else {
             Text(text = "Something went wrong...")
-        }
-    }
-}
-
-@OptIn(ExperimentalGlideComposeApi::class)
-@Composable
-fun PageImage(
-    imageUrl: String
-) {
-    // TODO: Add report
-    GlideImage(
-        modifier = Modifier.fillMaxWidth(),
-        model = imageUrl,
-        contentDescription = null,
-    ) {
-        it.placeholder(R.drawable.manga_cover)
-    }
-}
-
-@Composable
-fun NavigationField(
-    prevChapter: AggregateChapter?,
-    nextChapter: AggregateChapter?,
-    navigateToChapter: (chapterId: String) -> Unit
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        if (prevChapter != null) {
-            Button(
-                onClick = { navigateToChapter(prevChapter.id) },
-                colors = ButtonDefaults.textButtonColors()
-            ) {
-                Icon(
-                    imageVector = Icons.Default.ArrowBack,
-                    contentDescription = "Previous chapter"
-                )
-                Text(text = "Ch. ${prevChapter.chapter}")
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(1.dp))
-
-        if (nextChapter != null) {
-            Button(
-                onClick = { navigateToChapter(nextChapter.id) },
-                colors = ButtonDefaults.textButtonColors()
-            ) {
-                Text(text = "Ch. ${nextChapter.chapter}")
-                Icon(
-                    imageVector = Icons.Default.ArrowForward,
-                    contentDescription = "Next chapter"
-                )
-            }
         }
     }
 }
